webapp/tools: Adds server-log reset (via new StateDao.reset_in_instance()) to 'reset'
Former-commit-id: b260496b89f71437fe0592118118d7e1adef13ca

+++ webapp/tools.py
@@ -9,6 +9,7 @@ import sys
 import time
 
 from log_entry import LogEntry
+from state_dao import StateDao
 from ids.intrusion_classifier import IntrusionClassifier
 from ids.dir_utils import Dir, ModelDir
 import ids.ids_tools as ids_tools
@@ -307,18 +308,38 @@ def _init_file_handle(path):
 
 # pylint: disable-msg=W0613; (Unused argument)
 def reset_call(args):
-	""" Call _reset. Expects nothing. """
-	_reset()
+	""" Call _reset. Expects 'classifier' and 'server_log'. """
+	_reset(args.classifier, args.server_log)
 
 
-def _reset():
+def _reset(classifier, server_log):
 	""" Move the generated models to a sub-folder and reset the intrusion_classifier_history. """
 
-	print("Resetting...")
-	if os.path.lexists(_HISTORY_FILE):
-		os.remove(_HISTORY_FILE)
-		print("Removed the training history.")
-	message = ModelDir.reset_dir()
+	if not (classifier or server_log):
+		print("No selection was made! Add either --classifier or --server-log to reset.")
+		return
+
+	print("Resetting {}{}{}...".format(
+		"classifier" if classifier else "",
+		" and " if classifier and server_log else "",
+		"server log" if server_log else ""
+	))
+
+	message = ""
+
+	if classifier:
+		message += "\n# Classifier\n"
+		if os.path.lexists(_HISTORY_FILE):
+			os.remove(_HISTORY_FILE)
+			message += "Training history: File was removed.\n"
+		message += "Model directory: " + ModelDir.reset_dir()
+		if server_log:
+			message += "\n"
+
+	if server_log:
+		message += "\n# Server log\n"
+		message += StateDao.reset_in_instance()
+
 	print(message)
 
 
@@ -630,6 +651,8 @@ if __name__ == "__main__":
 		SPLIT_PARSER.set_defaults(function=split_call)
 
 		RESET_PARSER = SUBPARSERS.add_parser("reset", help="Reset the classifier")
+		RESET_PARSER.add_argument("--classifier", "-c", action="store_true")
+		RESET_PARSER.add_argument("--server-log", "-l", action="store_true")
 		RESET_PARSER.set_defaults(function=reset_call)
 
 		ANAL_PARSER = SUBPARSERS.add_parser("analyse", help="Analyse existing log data")
